requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langgraph>=0.2.50,<0.3.0
chromadb>=0.5.0,<0.6.0
dashscope>=1.14.0
pypdf>=4.0.0
python-docx>=1.1.0
jieba>=0.42.1
rank-bm25>=0.2.2
pydantic>=2.7.4,<3.0.0
requests>=2.31.0